backend/requirements.txt
Django==5.2.11
djangorestframework==3.16.1
django-cors-headers
djangorestframework-simplejwt
drf-spectacular
gunicorn
dj-database-url
psycopg[binary]
